done all functions but funcheck not yet

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: dshirais <dshirais@student.42vienna.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 16:04:13 by dshirais          #+#    #+#             */
-/*   Updated: 2026/01/02 23:12:44 by dshirais         ###   ########.fr       */
+/*   Updated: 2026/01/04 18:47:46 by dshirais         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ char	*make_result(char **save, char *buf, int re);
 int		eof_checker(char **buf, char **save);
 int		make_buf_and_read(char **buf, int fd);
 char	*make_save(char **save, char **buf, int re);
-size_t	check(const char *s, int flag);
+int	check(const char *s, int flag);
 char	*ft_strjoin_save(char *s1, char *s2);
 char	*get_next_line(int fd);
